pyproject.toml
[project]
name = "adk-agent-evals"
version = "0.1.0"
description = "Examples for ADK agent evals."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk==2.0.0a3",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "nbstripout>=0.9.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
    "ty>=0.0.26",
    "uv-build>=0.11.2",
]

[tool.ty.src]
include = ["src"]
exclude = ["tests/"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
main = "src.main:main"
eval = "src.main:main"

# RUFF section
[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "data",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
  "**/dist",
  "test_data",
  "tests",
]
# Lint and format also notebooks.
extend-include = ["**/*.ipynb"]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.14
target-version = "py314"

[tool.ruff.format]

preview = true

# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"


[tool.ruff.lint]
preview = true
select = [
  # pycodestyle
  "E",
  "W",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # pydocstyle
  "D",
  # mccabe
  "C90",
  # pep8 naming
  "N",
  # flake8 annotations
  "ANN",
  #flake8 bandit
  "S",
  # blind except
  "BLE",
  # flake8 boolean trap
  "FBT",
  # flake8 builtins
  "A",
  # flake8 commas
  "COM",
  # flake8 comprehensions
  "C4",
  # flake8-datetimez (DTZ)
  "DTZ",
  # flake8-implicit-str-concat (ISC)
  "ISC",
  # flake8-logging-format (G)
  "G",
  # flake8-no-pep420 (INP)
  "INP",
  # flake8-pie
  "PIE",
  # flake8-print
  "T20",
  # flake8-pyi
  "PYI",
  # flake8-pytest-style
  "PT",
  # flake8-quotes
  "Q",
  # flake8-raise
  "R",
  # flake8-return (RET)
  "RET",
  # flake8-self
  "SLF",
  # flake8-tidy-imports
  "TID",
  # flake8-type-checking
  "TCH",
  # flake8-gettext
  "INT",
  # flake8-unused-arguments
  "ARG",
  # flake8-use-pathlib
  "PTH",
  # eradicate
  "ERA",
  # pandas-vet
  "PD",
  # Pylint
  "PL",
  # tryceratops
  "TRY",
  # flynt
  "FLY",
  # NumPy-specific rules
  "NPY",
  # Airflow
  "AIR",
  # Perflint
  "PERF",
  # refurb
  "FURB",
  # flake8-logging
  "LOG",
  # Ruff-specific rules
  "RUF",
]
ignore = ["COM812", "ISC001", "S101"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"**/*.ipynb" = ["ANN", "D", "A004", "E402", "PTH", "E501", "T20"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

exclude = []

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.basedpyright]
include = ["src"]
typeCheckingMode = "off"
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "src/experimental",
  "src/typestubs",
  ".venv/",
]
ignore = ["src/oldstuff"]

reportMissingImports = true
reportMissingTypeStubs = false
reportCallIssue = false

pythonVersion = "3.14"

